Supporting cascaded profilers
Typically, some profilers must wait for other profilers to finish to start their compution.
An example is a semantic profiler which operates only on strings, or a statistical profiler which operates only on numbers.
Both of these profilers must wait for the data types to be inferred and operate on the unique values generated by such profiler.
To define profilers in a cascading fashion, simply modify config.js to list objects instead of strings for profiler spec.
The object should have 2 keys: module, used for the profiler name (loaded as node module).
The other key is cascade: recursively storing an array of profiler specs (either strings or objects)... etc.
An example config.js can be found in this commit.